pyproject.toml
[project]
name = "agent-edu-template-server"
version = "0.1.0"
description = "Agent Education Template Server"
requires-python = ">=3.11,<=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "opik>=1.10.4",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langchain-elasticsearch>=1.0.0",
    "pymupdf>=1.27.2.2",
    "tiktoken>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
